requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
shap
plotly
pillow
matplotlib
